feat: allow simplified trait object syntax
For the `#[derive(Inject)]` proc-macro, it's now possible to use
`#[provide(singleton = "dyn Foo")]` instead of having to spell out the
type it resolves to (e.g., `singleton = "Arc<dyn Foo>"`).

For transients, it will default to a `Box`, and for singletons to a
`ferrunix::Ref` (either `Arc`, or `Rc`).