fix: singleton behavior with derive macros
Previously, the derive macro wrapped all types, incorrectly, in a
`ferrunix::Ref` during registration and retrieval of types. This
incorrect wrapping is now removed since the constructor of a singleton
expect the concrete type, and this is also used as a key for retrieving
the singleton.

Trait objects (`dyn T`) are handled as an exception and are
automatically wrapped in a `Box`. This is, unfortunately, necessary as
the constructor returns a concrete, sized, type, and `dyn T` is unsized.